Rename map_view files to aligned_dict
